# server/controllers/Alpha_db/generated/EtudiantControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE EtudiantControllerGenerated.js PLEASE EDIT ../EtudiantController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import EtudiantModel from "../../../models/Alpha_db/EtudiantModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import EtudiantController from "../EtudiantController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/etudiant`;
    router.post(baseUrl + "", authorize([]), EtudiantController.create);
    router.delete(baseUrl + "/:id", authorize([]), EtudiantController.delete);
    router.get(baseUrl + "/findBy_classe/:key", authorize([]), EtudiantController.findBy_classe);
    router.get(baseUrl + "/findBy_matiere/:key", authorize([]), EtudiantController.findBy_matiere);
    router.get(baseUrl + "/:id", authorize([]), EtudiantController.get);
    router.get(baseUrl + "", authorize([]), EtudiantController.list);
    router.post(baseUrl + "/:id", authorize([]), EtudiantController.update);
  },


  // CRUD METHODS


  /**
  * EtudiantModel.create
  *   @description CRUD ACTION create
  *
  */
  create: async (req, res) => {
    try {
      const result = await EtudiantModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * EtudiantModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  delete: async (req, res) => {
    try {
      const result = await EtudiantModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * EtudiantModel.findBy_classe
  *   @description CRUD ACTION findBy_classe
  *   @param Objectid key Id of model to search for
  *
  */
  findBy_classe: async (req, res) => {
    try {
      const result = await EtudiantModel.findBy_classe(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * EtudiantModel.findBy_matiere
  *   @description CRUD ACTION findBy_matiere
  *   @param Objectid key Id of model to search for
  *
  */
  findBy_matiere: async (req, res) => {
    try {
      const result = await EtudiantModel.findBy_matiere(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * EtudiantModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  get: async (req, res) => {
    try {
      const result = await EtudiantModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * EtudiantModel.list
  *   @description CRUD ACTION list
  *
  */
  list: async (req, res) => {
    try {
      const result = await EtudiantModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * EtudiantModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  update: async (req, res) => {
    try {
      const result = await EtudiantModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;