# server/controllers/Alpha_db/generated/ExamenControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE ExamenControllerGenerated.js PLEASE EDIT ../ExamenController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import ExamenModel from "../../../models/Alpha_db/ExamenModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import ExamenController from "../ExamenController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/examen`;
    router.post(baseUrl + "", authorize([]), ExamenController.create);
    router.delete(baseUrl + "/:id", authorize([]), ExamenController.delete);
    router.get(baseUrl + "/findBy_classe/:key", authorize([]), ExamenController.findBy_classe);
    router.get(baseUrl + "/findBy_enseignant/:key", authorize([]), ExamenController.findBy_enseignant);
    router.get(baseUrl + "/findBy_etudiant/:key", authorize([]), ExamenController.findBy_etudiant);
    router.get(baseUrl + "/findBy_matiere/:key", authorize([]), ExamenController.findBy_matiere);
    router.get(baseUrl + "/:id", authorize([]), ExamenController.get);
    router.get(baseUrl + "", authorize([]), ExamenController.list);
    router.post(baseUrl + "/:id", authorize([]), ExamenController.update);
    router.post(baseUrl + "/:id/valide", ExamenController.valide);
  },


  // CRUD METHODS


  /**
  * ExamenModel.create
  *   @description CRUD ACTION create
  *
  */
  create: async (req, res) => {
    try {
      const result = await ExamenModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ExamenModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  delete: async (req, res) => {
    try {
      const result = await ExamenModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ExamenModel.findBy_classe
  *   @description CRUD ACTION findBy_classe
  *   @param Objectid key Id of model to search for
  *
  */
  findBy_classe: async (req, res) => {
    try {
      const result = await ExamenModel.findBy_classe(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ExamenModel.findBy_enseignant
  *   @description CRUD ACTION findBy_enseignant
  *   @param Objectid key Id of model to search for
  *
  */
  findBy_enseignant: async (req, res) => {
    try {
      const result = await ExamenModel.findBy_enseignant(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ExamenModel.findBy_etudiant
  *   @description CRUD ACTION findBy_etudiant
  *   @param Objectid key Id of model to search for
  *
  */
  findBy_etudiant: async (req, res) => {
    try {
      const result = await ExamenModel.findBy_etudiant(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ExamenModel.findBy_matiere
  *   @description CRUD ACTION findBy_matiere
  *   @param Objectid key Id of model to search for
  *
  */
  findBy_matiere: async (req, res) => {
    try {
      const result = await ExamenModel.findBy_matiere(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ExamenModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  get: async (req, res) => {
    try {
      const result = await ExamenModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * ExamenModel.list
  *   @description CRUD ACTION list
  *
  */
  list: async (req, res) => {
    try {
      const result = await ExamenModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * ExamenModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  update: async (req, res) => {
    try {
      const result = await ExamenModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
    
  
  // Custom APIs

  /**
  * ExamenModel.valide
  *   @description l&#x27;api est valide l&#x27;examen
  *   @param String id id de l&#x27;examen
  *   @returns Boolean
  *
  */
  valide: async (req, res) => {
    try {
      res.json({});
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },

};

export default generatedControllers;