# server/swagger.yaml
openapi: 3.0.0
info:
  title: Alpha
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /etudiant/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Etudiant
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Etudiant
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Etudiant"
    post:
      summary: CRUD ACTION update
      tags:
        - Etudiant
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Etudiant"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Etudiant"
  /etudiant/findBy_classe/{key}:
    get:
      summary: CRUD ACTION findBy_classe
      tags:
        - Etudiant
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Etudiant"
  /etudiant/findBy_matiere/{key}:
    get:
      summary: CRUD ACTION findBy_matiere
      tags:
        - Etudiant
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Etudiant"
    
  #-----  Examen  -----
...
  /examen/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Examen
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Examen
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Examen"
    post:
      summary: CRUD ACTION update
      tags:
        - Examen
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Examen"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Examen"
  /examen/findBy_classe/{key}:
    get:
      summary: CRUD ACTION findBy_classe
      tags:
        - Examen
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Examen"
  /examen/findBy_enseignant/{key}:
    get:
      summary: CRUD ACTION findBy_enseignant
      tags:
        - Examen
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Examen"
  /examen/findBy_etudiant/{key}:
    get:
      summary: CRUD ACTION findBy_etudiant
      tags:
        - Examen
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Examen"
  /examen/findBy_matiere/{key}:
    get:
      summary: CRUD ACTION findBy_matiere
      tags:
        - Examen
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Examen"
  /examen/{id}/valide:
    post:
      summary: l&#x27;api est valide l&#x27;examen
      security: []   # No security
      tags:
        - Examen
      parameters:
        - name: id
          in: path
          description: id de l&#x27;examen
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    
  #-----  Matiere  -----
...
  /matiere/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Matiere
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Matiere
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matiere"
    post:
      summary: CRUD ACTION update
      tags:
        - Matiere
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matiere"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Matiere"
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Etudiant:
      properties:
        _id:
          type: string
        date_de_naissance:
          type: integer
        nom:
          type: string
        prenom:
          type: string
      required:
        -  _id
        - prenom
    Examen:
      properties:
        _id:
          type: string
        note:
          type: number
        place:
          type: string
        valid:
          type: boolean
      required:
        -  _id
    Matiere:
      properties:
        _id:
          type: string
        nom:
          type: string
      required:
        -  _id
        - nom
...